Revert "Let lsp.client display completions provided by CompletionCollector"
This reverts commit bea46d4f58c560f9054dcdd362ad095c5f0f1367.
